hotfix : OrderDetail의 totalPrice 필드의 타입이 Integer로 변경되었습니다.

+++ src/main/java/dev/nbcsparta/assignment/commerce_backoffice/dto/OrderDetail.java
@@ -12,7 +12,7 @@ public record OrderDetail(
         String customerEmail,
         String productName,
         Integer quantity,
-        Long totalPrice,
+        Integer totalPrice,
         LocalDateTime orderDate,
         DeliveryStatus status,
         String ManagerName,
